It is now possible to search for a track using Spotify, however navigation is causing trouble so will have to rework it probbaly

+++ DjTracklistApp/Helpers/Extensions.swift
@@ -19,12 +19,12 @@ extension Int {
     /// - Parameter bpm: Beats Per Minute - for what bpm should the bars be calculated
     /// - Parameter timeUnit: The unit of time that self represents
     ///  - Parameter timeSignature: The time signature for which the bars amount should be calculated, is 4 in most songs -> 4 beats in a bar
-    func getBars(bpm: Double, timeUnit: TimeUnit, timeSignature: Int = 4) -> Int {
+    func getBars(bpm: Float, timeUnit: TimeUnit, timeSignature: Int = 4) -> Int {
         // Bars get calculated with the formula: Bars = (AmountOfMinutes * Beats Per Minute)/Time Signature
         switch (timeUnit) {
-        case .milliseconds: return Int((Double(self) * bpm) / (Double(timeSignature) * 60000.0))
-        case .seconds: return Int((Double(self) * bpm) / (Double(timeSignature) * 60.0))
-        case .minutes: return Int((Double(self) * bpm) / Double(timeSignature))
+        case .milliseconds: return Int((Double(self) * Double(bpm)) / (Double(timeSignature) * 60000.0))
+        case .seconds: return Int((Double(self) * Double(bpm)) / (Double(timeSignature) * 60.0))
+        case .minutes: return Int((Double(self) * Double(bpm)) / Double(timeSignature))
         }
         
     }
